+++ js/create-form/create-comments.js
@@ -1,4 +1,4 @@
-import {createComment} from './createFormFunctions/createComment.js';
+import {createComment} from './createFormFunctions/createComment.js.js';
 
 function createComments (cycles) {
   return Array.from({length: cycles}, createComment);
