+++ js/create-form/create-whole-form.js
@@ -1,4 +1,4 @@
-import { createFormObject } from './createFormObject.js';
+import { createFormObject } from './create-form-object.js';
 
 const similarFormObjects = Array.from({length: 25}, createFormObject);
