+++ js/form-objects-properties/create-avatar.js
@@ -1,4 +1,4 @@
-import { getRandomPositiveInteger } from './mathFunctions/randomNumber.js';
+import { getRandomPositiveInteger } from './mathFunctions/randomNumber.js.js';
 
 function createAvatar() {
   return `img/avatar-${getRandomPositiveInteger(1,6)}.svg`;
