+++ js/form-objects-properties/create-description.js
@@ -1,0 +1,6 @@
+import { getRandomItemFromArray } from './functionsForArray/getRandomItemFromArray.js.js';
+import { variantsOfDescriptions } from './variables/formVariables.js.js';
+
+const createDescription = () => getRandomItemFromArray(variantsOfDescriptions);
+
+export {createDescription};
