+++ js/form-objects-properties/create-message.js
@@ -1,0 +1,6 @@
+import { getRandomItemFromArray } from './functionsForArray/getRandomItemFromArray.js.js';
+import { variantsOfMessages } from './variables/formVariables.js.js';
+
+const createMessage = () => getRandomItemFromArray(variantsOfMessages);
+
+export {createMessage};
