+++ js/form-objects-properties/createUrl.js
@@ -1,4 +1,4 @@
-import {getRandomPositiveInteger} from './mathFunctions/randomNumber.js';
+import {getRandomPositiveInteger} from './mathFunctions/randomNumber.js.js';
 
 function createUrl() {
   return `photos/${getRandomPositiveInteger(1,25)}.jpg`;
