--- js/formObjectsPropertiesFunctions/createMessage.js
@@ -1,6 +1,0 @@
-import { getRandomItemFromArray } from './functionsForArray/getRandomItemFromArray.js';
-import { variantsOfMessages } from './variables/formVariables.js';
-
-const createMessage = () => getRandomItemFromArray(variantsOfMessages);
-
-export {createMessage};
